fix(app): moved file in correct folder
Moved file in correct folder
